requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.2.2=hbcca054_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
fonttools=4.51.0=pypi_0
gh=2.48.0=ha8f183a_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
matplotlib=3.8.4=pypi_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.26.4=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py310h06a4308_0
pyparsing=3.1.2=pypi_0
python=3.10.14=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
readline=8.2=h5eee18b_0
scipy=1.13.0=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
tk=8.6.12=h1ccaba5_0
tzdata=2024.1=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
